requirements.txt
altair==4.1.0
astor==0.8.1
astroid @ file:///tmp/build/80754af9/astroid_1592495881661/work
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
boto3==1.14.7
botocore==1.17.7
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
idna==2.9
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython @ file:///tmp/build/80754af9/ipython_1592348411427/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.1
olefile==0.46
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pydeck==0.4.0b2
Pygments==2.6.1
pylint @ file:///tmp/build/80754af9/pylint_1592496044040/work
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.15.0
streamlit==0.62.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
torch==1.5.1
torchvision==0.6.0a0+35d732a
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
watchdog==0.10.2
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
